Fix matching only working for imap in Thunderbird.
Pull request #334 introduces excaping regex character in order to get matching
working with some translations. However, it failed to also escape regex chars
in the placeholder strings, so it broke everything except for imap (because imap
is the only protocol that does not use '$' in the placeholder).

This patch uses the same escaping on the placeholder strings, so now it works
again for all protocols.

Fixes #351.