Add support for Google Calendar Provider extension in Thunderbird.
This add a XUL overlay for the authentication dialog provided by the "Provider
for Google Calendar" extension in Thunderbird. The dialog uses an embedded
browser to load a web page, but the embedded browser does not provide the usual
DOM events, so the existing KeeFox web form filler does not work. So instead,
this treats it more like commonDialog that handles XUL dialogs. In fact, most
of the code is copied from commonDialog and modified to work with the gdata-provider
dialog.

Fixes luckyrat/KeeFox#375