Use global setTimeout to prevent race condition when dialog is opened.
It is possilble that during application startup, a dialog can open in the 50ms
between when these timers are set and the callback. It is probably longer than
50ms since the application is busy during startup, which increases the chance
of a problem.

Since we are bumping the minimum version to 22 (#356), we can now use the
setTimout from https://developer.mozilla.org/en-US/docs/Mozilla/JavaScript_code_modules/Timer.jsm

Since the timeout is no longer tied to window, it is free to callback even though
a modal dialog box opens.

Fixes #171