Restore default "signon.rememberSignons" preference when KeeFox is disabled/uninstalled.
Instead of trying to restore the users previous selection, the sets the preference back
to the default value so that the user gets an out-of-box experience. In both Firefox
and Thuderbird, the user will be asked if they want to use the built-in password manager
the first time they enter a password after restart. This is better than the alternative
of having to dig into the preferences and turn the feature back on (particularly in
Thunderbird where it is an advanced preference only availible in the config editor).

Fixes luckyrat/KeeFox#389