Update gdata-provider.js for changes in google oauth
Google has changed their oauth so that username and password are on 2 different
pages. This modifies the gdata-provider code so that each page is handled without
crashing anymore. Also fixes autosubmitting for the dialog. User will still have
to click allow/deny at the end (3rd page), but "next" and "signIn" are automatically
clicked on the first two pages if autosubmit is enabled.

Fixes luckyrat/KeeFox#642