Fix commonDialog string matching for some transations in Firefox 50+
As part of the "fix" for https://bugzilla.mozilla.org/show_bug.cgi?id=1277895,
Firefox is stripping newline characters from localizations before displaying
the text. So, this copies the newline stripping code so that we are matching
what Firefox is actuall displaying rather than what is in the string resource.

Fixes #736